hero block in progress changes
