--- blocks/heroblock/heroblock.css
@@ -1,45 +1,0 @@
-.hero-container .hero-wrapper {
-  max-width: unset;
-  padding: 0;
-}
-
-.hero-block-image {
-  height: 400px;
-  overflow: hidden;
-}
-
-.hero-block-title {
-  font-size: 42px;
-  margin-bottom: 30px;
-}
-
-.hero-block-caption {
-  width: 85%;
-  background: #fff;
-  padding: 30px;
-  border-top-right-radius: 20px;
-  margin-top: -60px;
-  position: relative;
-  font-size: 22px;
-}
-
-.hero-block-caption p,
-.hero-block-intro-text p {
-  margin: 0
-}
-
-.hero-block-intro-text {
-  font-size: 16px;
-  width: 85%;
-}
-
-.hero-block-image img {
-  width: 100%;
-  border-top-right-radius: 50px;
-}
-
-@media (width >= 900px) {
-  .hero {
-    padding: 40px 32px;
-  }
-}
