+++ component-definition.json
@@ -171,21 +171,6 @@
               }
             }
           }
-        },
-        {
-          "title": "HeroBlock",
-          "id": "heroblock",
-          "plugins": {
-            "xwalk": {
-              "page": {
-                "resourceType": "core/franklin/components/block/v1/block",
-                "template": {
-                  "name": "HeroBlock",
-                  "model": "heroblock"
-                }
-              }
-            }
-          }
         }
       ]
     }
