+++ component-filters.json
@@ -15,9 +15,7 @@
       "hero",
       "cards",
       "columns",
-      "fragment",
-      "heroblock",
-      "assetselector"
+      "fragment"
     ]
   },
   {
