+++ component-models.json
@@ -212,75 +212,6 @@
         "valueType": "string",
         "name": "imageAlt",
         "label": "Alternative text for image",
-        "value": ""
-      },
-      {
-        "component": "text",
-        "valueType": "string",
-        "name": "imageLocation",
-        "label": "Image Location/Caption",
-        "value": ""
-      },
-      {
-        "component": "tab",
-        "label": "Text",
-        "name": "text"
-      },
-      {
-        "component": "text",
-        "valueType": "string",
-        "name": "title",
-        "label": "Title",
-        "value": ""
-      },
-      {
-        "component": "richtext",
-        "valueType": "string",
-        "name": "introText",
-        "label": "Intro Text",
-        "value": ""
-      },
-      {
-        "component": "tab",
-        "label": "Logo",
-        "name": "logo"
-      },
-      {
-        "component": "reference",
-        "valueType": "string",
-        "name": "logoImage",
-        "label": "Logo Image",
-        "multi": false
-      },
-      {
-        "component": "text",
-        "valueType": "string",
-        "name": "logoAlt",
-        "label": "Logo Alt",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "id": "heroblock",
-    "fields": [
-      {
-        "component": "tab",
-        "label": "Asset",
-        "name": "asset"
-      },
-      {
-        "component": "reference",
-        "valueType": "string",
-        "name": "image",
-        "label": "Image",
-        "multi": false
-      },
-      {
-        "component": "text",
-        "valueType": "string",
-        "name": "imageAlt",
-        "label": "Alternative text for image",
         "description": "Write descriptive, concise, relevant alt text to describe the image in context, enhancing accessibility forvisually impaired users and improving SEO.",
         "value": "",
         "required": true
